+++ Source/SelfService/Web/integrations/connection/connectionDetails/consumeDataRestAPI/Credentials/credentialsTableGrid/DeleteCredentialsDialog.tsx
@@ -2,7 +2,9 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 import React, { Dispatch } from 'react';
+
 import { AlertDialog } from '@dolittle/design-system/';
+
 import { Box, Typography } from '@mui/material';
 
 export type DeleteCredentialsDialogState = {
@@ -37,22 +39,23 @@ export type DeleteCredentialsDialogProps = {
 };
 
 export const DeleteCredentialsDialog = ({ dispatch, state, onDelete }: DeleteCredentialsDialogProps) => {
-
     const hasMany = state.credentials.length > 1;
 
-    return <AlertDialog
-        id='delete-credentials-dialog'
-        title={`Delete credential${hasMany ? 's' : ''}`}
-        description={`Are you sure you want to delete the selected credential${hasMany ? 's' : ''}? This action cannot be undone and will impact any applications currently using these credentials.`}
-        confirmBtnText='Delete'
-        confirmBtnColor='error'
-        isOpen={state.open}
-        onConfirm={() => onDelete(state.credentials)}
-        onCancel={() => dispatch({ type: 'close' })}
-        isLoading={state.isLoading}
-    >
-        <Box marginTop={2}>
-            {state.credentials.map((serviceAccount) => <Typography key={serviceAccount}>{`"${serviceAccount}"`}</Typography>)}
-        </Box>
-    </AlertDialog>;
+    return (
+        <AlertDialog
+            id='delete-credentials-dialog'
+            title={`Delete credential${hasMany ? 's' : ''}`}
+            description={`Are you sure you want to delete the selected credential${hasMany ? 's' : ''}? This action cannot be undone and will impact any applications currently using these credentials.`}
+            confirmBtnText='Delete'
+            confirmBtnColor='error'
+            isOpen={state.open}
+            onConfirm={() => onDelete(state.credentials)}
+            onCancel={() => dispatch({ type: 'close' })}
+            isLoading={state.isLoading}
+        >
+            <Box sx={{ mt: 2 }}>
+                {state.credentials.map(serviceAccount => <Typography key={serviceAccount}>{`"${serviceAccount}"`}</Typography>)}
+            </Box>
+        </AlertDialog>
+    );
 };
